backend/requirements/base.txt
amqp == 1.4.4
anyjson == 0.3.3
argh == 0.24.1
billiard == 3.3.0.16
bleach == 1.4
boto == 2.27.0
celery == 3.1.9
dj-database-url == 0.3.0
Django == 1.6.5
django-allauth == 0.15.0
django-bootstrap3 == 3.0.0
django-braces == 1.3.1
django-celery == 3.1.9
django-debug-toolbar == 1.2.1
django-extensions == 1.3.3
django-fsm == 1.5.1
django-model-utils == 2.0.2
django-object-actions == 0.4.0
django-redis == 3.6.1
django-storages == 1.1.8
django-suit == 0.2.7
django-vanilla-views == 1.0.2
django-waffle == 0.10
docopt == 0.4.0
envdir == 0.6.1
feedparser == 5.1.3
html5lib == 0.999
ipython == 2.0.0
Jinja2 == 2.7.2
kombu == 3.0.13
logutils == 0.3.3
mailchimp == 2.0.7
meld3 == 0.6.10
oauthlib == 0.6.1
pathtools == 0.1.2
psycopg2 == 2.5.2
Pygments == 1.6
python-openid == 2.2.5
pytz == 2013.9
PyYAML == 3.10
pyzmq == 14.1.1
raven == 4.0.4
redis == 2.9.1
requests == 2.2.1
requests-oauthlib == 0.4.0
setuptools == 3.4.3
South == 0.8.4
supervisor == 3.0
tornado == 3.2
Unipath == 1.0
watchdog == 0.7.1
django-supervisor == 0.3.2
